fix scheduler keys

+++ novawallet/Common/Substrate/Types/Referenda/Referenda.swift
@@ -2,5 +2,5 @@ import Foundation
 
 enum Referenda {
     typealias TrackId = UInt16
-    typealias ReferendumIndex = UInt64
+    typealias ReferendumIndex = UInt32
 }
